区别对待二进制与源码方式代码库上传
在以git clone方式得到的本地代码库中，上传代码库需要重新构建。
这种方式可以方便程序调试，例如某些试验性的改动并不需要提交到
代码库中。

在以二进制方式下载的到的本地代码库中，上传代码库不需要重新构
建，仅重新部署即可。

+++ broker/apps.go
@@ -6,7 +6,6 @@ import (
 	"crypto/rand"
 	"crypto/sha1"
 	"encoding/hex"
-	"encoding/json"
 	errs "errors"
 	"fmt"
 	"io"
@@ -21,7 +20,6 @@ import (
 	"github.com/Sirupsen/logrus"
 	"golang.org/x/net/context"
 
-	"github.com/cloudway/platform/api/types"
 	"github.com/cloudway/platform/auth/userdb"
 	"github.com/cloudway/platform/config/defaults"
 	"github.com/cloudway/platform/container"
@@ -29,6 +27,7 @@ import (
 	"github.com/cloudway/platform/pkg/errors"
 	"github.com/cloudway/platform/pkg/files"
 	"github.com/cloudway/platform/pkg/manifest"
+	"github.com/cloudway/platform/pkg/serverlog"
 	"github.com/cloudway/platform/scm"
 )
 
@@ -197,46 +196,11 @@ func populateFromTemplate(scm scm.SCM, opts *container.CreateOptions, template s
 	return err
 }
 
-type logWriter struct {
-	out io.Writer
-	enc *json.Encoder
-}
-
-func newLogWriter(out io.Writer) *logWriter {
-	return &logWriter{
-		out: out,
-		enc: json.NewEncoder(out),
-	}
-}
-
-func (log *logWriter) Write(p []byte) (n int, err error) {
-	stream := types.ServerLog{Message: string(p)}
-	if err = log.enc.Encode(&stream); err != nil {
-		return 0, err
-	}
-
-	type Flusher interface {
-		Flush()
-	}
-
-	type ErrFlusher interface {
-		Flush() error
-	}
-
-	switch b := log.out.(type) {
-	case Flusher:
-		b.Flush()
-	case ErrFlusher:
-		err = b.Flush()
-	}
-	return len(p), err
-}
-
 func deployRepo(scm scm.SCM, opts *container.CreateOptions, containers []*container.Container) error {
 	if opts.Log == nil {
 		return scm.Deploy(opts.Namespace, opts.Name, "")
 	} else {
-		var log = newLogWriter(opts.Log)
+		var log = serverlog.NewLogWriter(opts.Log)
 		return scm.DeployWithLog(opts.Namespace, opts.Name, "", log, log)
 	}
 }
@@ -679,30 +643,35 @@ func (br *UserBroker) Download(name string) (io.ReadCloser, error) {
 }
 
 // Upload application repository from a archive file.
-func (br *UserBroker) Upload(name string, content io.Reader) error {
-	repodir, err := archive.PrepareRepo(content, false)
-	if repodir != "" {
-		defer os.RemoveAll(repodir)
-	}
-	if err != nil {
-		return err
-	}
-
-	// deploy to containers
-	containers, err := br.FindApplications(br.ctx, name, br.Namespace())
-	if err != nil {
-		return err
-	}
-	if len(containers) == 0 {
-		return ApplicationNotFoundError(name)
-	}
-	for _, c := range containers {
-		er := c.Deploy(br.ctx, repodir)
-		if er != nil {
-			err = er
+func (br *UserBroker) Upload(name string, content io.Reader, binary bool, logger io.Writer) error {
+	if binary {
+		repodir, err := archive.PrepareRepo(content, false)
+		if repodir != "" {
+			defer os.RemoveAll(repodir)
 		}
+		if err != nil {
+			return err
+		}
+
+		// deploy to containers
+		containers, err := br.FindApplications(br.ctx, name, br.Namespace())
+		if err != nil {
+			return err
+		}
+		if len(containers) == 0 {
+			return ApplicationNotFoundError(name)
+		}
+		for _, c := range containers {
+			er := c.Deploy(br.ctx, repodir)
+			if er != nil {
+				err = er
+			}
+		}
+		return err
+	} else {
+		// deploy the repository
+		return scm.DeployRepository(br.DockerClient, br.ctx, name, br.Namespace(), content, logger, logger)
 	}
-	return err
 }
 
 func (br *UserBroker) Dump(name string) (io.ReadCloser, error) {
